Fix CI concurrency issues
As you mentioned, having multiple clients open will nullify the pooling work. In CI, we can limit the test concurrency to work around this.

In the HTTP client, we can better deserialize the server exceptions so that these exceptions are raised in a way they can be caught and matched better.